feat(admin): 新增管理员头像上传功能 feat(user): 新增用户头像上传功能 fix: 修正adminInfoURL调用错误 fix: 修正getColumnValues函数逻辑 feat(search): 添加搜索功能接口与实现 fix: 修正registrationInfo接口调用路径 fix: 修正registrationStatus接口返回值处理 fix: 修正导航栏显示逻辑 fix: 修正查询招聘信息逻辑 fix: 修正查询公司信息逻辑 fix: 修正用户头像上传逻辑 fix: 修正管理员头像上传逻辑 style: 更新导航栏样式与功能按钮样式

+++ recruit_services/src/main/java/com/queryx/recruiting_website/service/impl/UserServiceImpl.java
@@ -33,7 +33,7 @@ public class UserServiceImpl implements UserService {
     @Autowired
     private TDResumeMapper resumeMapper;
 
-    @Value("${file.upload-path-avatar")
+    @Value("${file.upload-path-avatar}")
     private String uploadPathAvatar;
 
     public static final String PHONE = "(^1[3-9]\\d{9}$)";
@@ -93,7 +93,7 @@ public class UserServiceImpl implements UserService {
             return CommonResp.fail(AppHttpCodeEnum.SYSTEM_ERROR, null);
         }
         // 有 -> 删除
-        if (user.getUserAvatar() != null || user.getUserAvatar().trim().isEmpty()) {
+        if (user.getUserAvatar() != null && user.getUserAvatar().trim().isEmpty()) {
             File oldFile = new File(uploadPathAvatar + user.getUserAvatar());
             if (!(oldFile.exists() && oldFile.delete())) {
                 return CommonResp.fail(AppHttpCodeEnum.AVATAR_DELETE_ERROR, null);
@@ -110,15 +110,12 @@ public class UserServiceImpl implements UserService {
                 return CommonResp.fail(AppHttpCodeEnum.AVATAR_UPLOAD_ERROR, null);
             }
             // 更新数据库
-            path = "/" + Common.getUploadFolderName(uploadPathAvatar, "/", newFileName);
+            path = "/" + Common.getLastPath(uploadPathAvatar, "/", newFileName);
             user.setUserAvatar(path);
             userMapper.updateById(user);
         } catch (Exception e) {
             return CommonResp.fail(AppHttpCodeEnum.AVATAR_UPLOAD_ERROR, null);
         }
-//        String[] strings = uploadPathAvatar.split("/");
-//        String uploadFolder = strings[strings.length - 1];
-//        return CommonResp.success(uploadFolder + "/" + newFileName);
         return CommonResp.success(Common.getImgURL() + path);
     }
